backend/requirements.txt
fastapi
websockets
python-dotenv
google-generativeai
google-cloud-firestore
uvicorn
praw
